--- a/JobScheduler.Benchmarks/ManyJobsBenchmark.cs
+++ b/JobScheduler.Benchmarks/ManyJobsBenchmark.cs
@@ -11,19 +11,23 @@ public class ManyJobsBenchmark
     /// <summary>
     /// The thread count tested
     /// </summary>
-    [Params(0)] public int Threads = 0; 
-    
+    [Params(0)] public int Threads = 0;     
 
     /// <summary>
     /// Whether the benchmark should do a dry run before starting the benchmark, to fill up any data structures that might allocate
     /// on frame 1.
     /// </summary>
-    [Params(true, false)] public bool SetupJobsFirst;
+    [Params(true)] public bool ExcludeFirstWave;
 
     /// <summary>
-    /// The amount of jobs to run.
+    /// The maximum amount of concurrent jobs active at one time.
     /// </summary>
-    [Params(10, 100, 1000000)] public int JobCount;
+    [Params(128)] public int ConcurrentJobs;
+
+    /// <summary>
+    /// How many sequences of <see cref="ConcurrentJobs"/> to run.
+    /// </summary>
+    [Params(10240)] public int Waves;
 
     JobHandle[] Handles = null!;
     private class EmptyJob : IJob
@@ -31,26 +35,27 @@ public class ManyJobsBenchmark
         public void Execute() { }
     }
 
-    private static IJob Empty => new EmptyJob();
+    private readonly static EmptyJob Empty = new();
 
     [IterationSetup]
     public void Setup()
     {
         Scheduler = new(nameof(ManyJobsBenchmark), Threads);
-        Handles = new JobHandle[JobCount];
+        Handles = new JobHandle[ConcurrentJobs];
 
-        if (SetupJobsFirst)
+        if (ExcludeFirstWave)
         {
-            for (int i = 0; i < JobCount; i++)
+            Waves--;
+            for (int i = 0; i < ConcurrentJobs; i++)
             {
                 Handles[i] = Scheduler.Schedule(Empty);
             }
             Scheduler.Flush();
-            for (int i = 0; i < JobCount; i++)
+            for (int i = 0; i < ConcurrentJobs; i++)
             {
                 Handles[i].Complete();
             }
-            for (int i = 0; i < JobCount; i++)
+            for (int i = 0; i < ConcurrentJobs; i++)
             {
                 Handles[i].Return();
             }
@@ -66,18 +71,21 @@ public class ManyJobsBenchmark
     [Benchmark]
     public void BenchmarkParallel()
     {
-        for (int i = 0; i < JobCount; i++)
+        for (int w = 0; w < Waves; w++)
         {
-            Handles[i] = Scheduler.Schedule(Empty);
-        }
-        Scheduler.Flush();
-        for (int i = 0; i < JobCount; i++)
-        {
-            Handles[i].Complete();
-        }
-        for (int i = 0; i < JobCount; i++)
-        {
-            Handles[i].Return();
+            for (int i = 0; i < ConcurrentJobs; i++)
+            {
+                Handles[i] = Scheduler.Schedule(Empty);
+            }
+            Scheduler.Flush();
+            for (int i = 0; i < ConcurrentJobs; i++)
+            {
+                Handles[i].Complete();
+            }
+            for (int i = 0; i < ConcurrentJobs; i++)
+            {
+                Handles[i].Return();
+            }
         }
     }
 }