Adds a benchmark framework with an initial benchmark, based of Arch's solution.

Current results as of this PR: 
```
BenchmarkDotNet v0.13.9+228a464e8be6c580ad9408e98f18813f6407fb5a, Windows 10 (10.0.19045.3448/22H2/2022Update)
AMD Ryzen 7 5800X3D, 1 CPU, 16 logical and 8 physical cores
.NET SDK 7.0.401
  [Host]     : .NET 7.0.11 (7.0.1123.42427), X64 RyuJIT AVX2
  Job-ZPCEAS : .NET 7.0.11 (7.0.1123.42427), X64 RyuJIT AVX2

InvocationCount=1  UnrollFactor=1  
```
| Method              | Threads | ExcludeFirstWave | ConcurrentJobs | Waves | Mean        | Error    | StdDev   | Gen0      | Gen1      | Allocated  |
|-------------------- |-------- |----------------- |--------------- |------ |------------:|---------:|---------:|----------:|----------:|-----------:|
| **BenchmarkParallel**   | **0**       | **True**             | **32**             | **10240** |    **343.2 ms** |  **4.13 ms** |  **3.66 ms** |         **-** |         **-** |      **600 B** |
| BenchmarkSequential | 0       | True             | 32             | 10240 |  3,732.7 ms | 17.31 ms | 16.19 ms |         - |         - |      600 B |
| **BenchmarkParallel**   | **0**       | **True**             | **128**            | **10240** |  **1,344.9 ms** | **11.85 ms** | **11.09 ms** | **1000.0000** | **1000.0000** | **54921816 B** |
| BenchmarkSequential | 0       | True             | 128            | 10240 | 14,916.6 ms | 97.26 ms | 81.22 ms |         - |         - |      600 B |


This will be easier to address prior to #7.